backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
langchain==0.3.7
langchain-google-genai==2.0.7
langchain-community==0.3.7
faiss-cpu==1.8.0
pymupdf==1.24.10
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
pydantic==2.9.2
